Fix namespace-scoped Grafana links
The `/namespaces` page in the web dashboard was rendering broken Grafana
links, containing an extra `var-namespace=` param, for example:
```
/grafana/dashboard/db/linkerd-namespace?var-namespace=&var-namespace=emojivoto
```

Root cause was the `GrafanaLink` component taking both `resource` and
`namespace` properties, but not special-casing when
`resource === 'namespace' && namespace === ''`.

Modify the `GrafanaLink` component to omit the `var-namespace` param
when a `namespace` property is not provided.

Signed-off-by: Andrew Seigner <siggy@buoyant.io>