WIP: add functionality to deal with windows nodes
WIP:  #2882

*Subject*
Proxies can be deployed on nodes that run a windows image, but they stop their working.

*Soultion*
addind a function `checkNodeOS` that checks if 'OS' taints have been applied to the nodes or if the NodeSelector contains the word windows.

*Current Functionality*
- Logs an info message when a windows node is scheduled
- skips injection on windows nodes

TO-DO:
- add tests
- add doccumentation for taints & tolerations

To validate the changes, I have used the file described in https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/

Signed-off-by: Vani Singh <vanisingh@live.co.uk>