Fix SC1090 shellcheck issues in shell script files
The SC1090 "Can't follow non-constant source" issue is addressed in the
way suggested in shellcheck's documentation; the source paths are
pointed out in shellcheck comments. By adding the bin dir to the -P
shellcheck CLI parameter, we avoid having to state the bin directory in
each and every script file.

Change-Id: If22fa0313ea435de4fbeb1f807850ab6a80a392a
Signed-off-by: Joakim Roubert <joakimr@axis.com>