Introduce new env var in proxy partial for service topology (#4498)
For service topology, the destination service needs to know the topology of the source of the traffic
in order to set the priority for endpoints accordingly. By introducing a new env variable in the
proxy partial tmpl, the proxy can read it and include the source of the node as part of its service
discovery calls.

LINKERD2_PROXY_SOURCE_NODE is set by using the downward API (value taken from
spec.nodeName).

Fix #4498

Signed-off-by: Matei David <matei.david.35@gmail.com>