Add a Windows CLI Test
This PR adds a new cli test to see if installation yamls are correctly
generated even on windows, this is important because of all the file
path difference between windows and Linux, and if any code uses a wrong
format might cause the chart generation commands to fail on windows.

This creates a separate workflow for both release, and PR tests, and
downloads the windows binary and runs the test.

Also, all the exisiting integration tests are moved in to
/tests/integration to separate from /test/cli as this test does not fall
under integration tests category

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>