Make uncheduble pods check warning only (#4675)
Currently commands that need access to the public api are executing the `LinkerdControlPlaneExistenceChecks` This set of checks includes one that specifically checks that there is no unscheduable pods. In fact in order to run commands like stat and edge we do not need to meet that requirement.

This change relaxes all this by makind the no unschedulable pods a warning only check. Fixes #3940

Signed-off-by: Zahari Dichev zaharidichev@gmail.com