check: make `checkOutput` public for conformance validation
Problem

The conformance validation project needs a way to validate the output
of `linkerd check`. A very simple way to do this would be to capture the
output of `check` and compare it against some predefined `.golden` file
which holds the desired output, much like the integration tests. The
issue with this is that for conformance validation, users are allowed to
specify the version of Linkerd CLI to use for performing the tests, and
it is possible for different versions to have different sets of checks.
It is also not feasible to have a `.golden` file for each stable and
edge version.

Solution

Capture the JSON output of the check results (`linkerd check -o json`), and
unmarshal it to an instance of `CheckOutput`. We can now iterate through
this object to see if any `Check.Result` == "error".

Signed-off-by: Mayank Shah <mayankshah1614@gmail.com>